Sync'ed with the master and address review comments.

+++ hadoop-hdds/container-service/src/main/java/org/apache/hadoop/ozone/container/keyvalue/interfaces/ChunkManager.java
@@ -119,10 +119,10 @@ public interface ChunkManager {
 
   static int getBufferCapacityForChunkRead(ChunkInfo chunkInfo,
       int defaultReadBufferCapacity) {
-    int bufferCapacity = 0;
+    long bufferCapacity = 0;
     if (chunkInfo.isReadDataIntoSingleBuffer()) {
       // Older client - read all chunk data into one single buffer.
-      bufferCapacity = (int) chunkInfo.getLen();
+      bufferCapacity = chunkInfo.getLen();
     } else {
       // Set buffer capacity to checksum boundary size so that each buffer
       // corresponds to one checksum. If checksum is NONE, then set buffer
@@ -140,7 +140,7 @@ public interface ChunkManager {
     }
     // If the buffer capacity is 0, set all the data into one ByteBuffer
     if (bufferCapacity == 0) {
-      bufferCapacity = (int) chunkInfo.getLen();
+      bufferCapacity = chunkInfo.getLen();
     }
 
     if (bufferCapacity > Integer.MAX_VALUE) {
